+++ package.json
@@ -212,6 +212,7 @@
         "axios": "0.18.0",
         "chalk": "2.4.1",
         "electron-store": "2.0.0",
+        "lodash.isequal": "^4.5.0",
         "mkdirp": "^0.5.1",
         "semver": "^5.4.1",
         "web3": "^1.0.0-beta.34",
