+++ src/components/tasks/TaskItem.js
@@ -7,21 +7,11 @@ import {Tooltip} from 'react-tippy';
 
 import { Motion, spring } from 'react-motion';
 import { convertSecsToHMS, timeStampToHR } from './../../utils/secsToHMS'
+import {taskStatus} from './../../constants/statusDicts'
 
 import * as Actions from '../../actions'
 
 import Preview from './Preview'
-
-const status = Object.freeze({
-    WAITINGFORPEER: 'Waiting for peer',
-    NOTREADY: 'Not started',
-    READY: 'Ready',
-    WAITING: 'Waiting',
-    COMPUTING: 'Computing',
-    FINISHED: 'Finished',
-    TIMEOUT: 'Timeout',
-    RESTART: 'Restart'
-})
 
 const ETH_DENOM = 10 ** 18;
 
@@ -53,6 +43,7 @@ export class TaskItem extends React.Component {
                 actions.fetchSubtasksList(item.id)
                 return interval
             }
+        if(item.status === taskStatus.COMPUTING)
             this.liveSubList = setInterval(interval(), 5000)
     }
 
@@ -83,7 +74,7 @@ export class TaskItem extends React.Component {
         const {subtasksList} = this.props
 
         switch (item.status) {
-        case status.TIMEOUT:
+        case taskStatus.TIMEOUT:
             return <div>
             	<span>Task time: {timeStampToHR((item.last_updated - item.time_started), true)}</span>
             	<span className="bumper"> | </span>
@@ -91,28 +82,28 @@ export class TaskItem extends React.Component {
             	<span>{timeStampToHR(item.last_updated)}</span>
             </div>
 
-        case status.NOTREADY:
+        case taskStatus.NOTREADY:
             return <div>
                 <span>Duration: {convertSecsToHMS(item.duration)}</span>
                 <span className="bumper"> | </span>
                 <span className="duration--preparing">Preparing for computation... </span>
             </div>
 
-        case status.WAITING:
+        case taskStatus.WAITING:
             return <div>
                 <span>Duration: {convertSecsToHMS(item.duration)}</span>
                 <span className="bumper"> | </span>
                 <span className="duration--preparing">Waiting for computation... </span>
             </div>
 
-        case status.RESTART:
+        case taskStatus.RESTART:
             return <div>
                 <span>Task time: {timeStampToHR((item.last_updated - item.time_started), true)}</span>
                 <span className="bumper"> | </span>
                 <span className="duration--restarted">Restarted</span>
             </div>
 
-        case status.COMPUTING:
+        case taskStatus.COMPUTING:
             return <div>
                 <span>Duration: {convertSecsToHMS(item.duration)}</span>
                 <span className="bumper"> | </span>
@@ -170,7 +161,7 @@ export class TaskItem extends React.Component {
                                 {this._fetchStatus(item)}
                                 <div className="info__task">
                                     <div>
-                                        <span>Frames: {(options && options.frames) || 0}</span>
+                                        <span>Frames: {(options && options.frame_count) || 0}</span>
                                         <span className="bumper"> | </span>
                                         <span> Resolution: {(options && options.resolution.join("x")) || 0}</span>
                                         <span className="bumper"> | </span>
@@ -215,17 +206,17 @@ export class TaskItem extends React.Component {
                           html={<p>Restart</p>}
                           position="right"
                           trigger="mouseenter"
-                          disabled={item.status !== status.TIMEOUT}>
+                          disabled={item.status !== taskStatus.TIMEOUT}>
                             <span 
                                 className="icon-progress-clockwise" 
                                 tabIndex="0" 
                                 aria-label="Restart Task" 
-                                onClick={(item.status == status.TIMEOUT 
-                                         || item.status == status.FINISHED) 
+                                onClick={(item.status == taskStatus.TIMEOUT 
+                                         || item.status == taskStatus.FINISHED) 
                                             ? _handleRestartModal 
                                             : undefined} 
-                                disabled={!(item.status === status.TIMEOUT
-                                          || item.status === status.FINISHED)}></span>
+                                disabled={!(item.status === taskStatus.TIMEOUT
+                                          || item.status === taskStatus.FINISHED)}></span>
                         </Tooltip>
                         <Tooltip
                           html={<p>Delete</p>}
