+++ src/constants/statusDicts.js
@@ -1,0 +1,21 @@
+const testStatusDict = Object.freeze({
+    STARTED: 'Started',
+    SUCCESS: 'Success',
+    ERROR: 'Error'
+})
+
+const taskStatus = Object.freeze({
+    WAITINGFORPEER: 'Waiting for peer',
+    NOTREADY: 'Not started',
+    READY: 'Ready',
+    WAITING: 'Waiting',
+    COMPUTING: 'Computing',
+    FINISHED: 'Finished',
+    TIMEOUT: 'Timeout',
+    RESTART: 'Restart'
+})
+
+module.exports = {
+    testStatusDict,
+    taskStatus
+}
