+++ src/scss/main.scss
@@ -243,6 +243,13 @@ select {
        
         }
     }
+
+    &--small{
+        width: 50px;
+        height: 30px;
+        padding: 0 2px;
+        font-size: 8pt;
+    }
 }
 
 input[type=text],
